+++ controller/teachersController.js
@@ -38,38 +38,55 @@ exports.getTeacherById = (req, res, next) => {
 
 exports.insertTeacher = (req, res, next) => {
   let object = new teacherSchema(req.body);
-  object.image = `${Date.now()}-${Math.random()}-${req.file.originalname}`;
-  fs.writeFile(`./photos/teacher/${object.image}`, req.file.buffer, (err) => {
-    if (err) return next(err);
+  if (req.file) {
+    object.image = `${Date.now()}-${Math.random()}-${req.file.originalname}`;
+    fs.writeFile(`./photos/teacher/${object.image}`, req.file.buffer, (err) => {
+      if (err) return next(err);
+      object
+        .save()
+        .then((data) => {
+          res.status(200).json({ data });
+        })
+        .catch((error) => next(error));
+    });
+  } else {
     object
       .save()
       .then((data) => {
         res.status(200).json({ data });
       })
       .catch((error) => next(error));
-  })
+  }
 };
 
 exports.updateTeacher = (req, res, next) => {
-  req.body.image = `${Date.now()}-${Math.random()}-${req.file.originalname}`;
-  fs.writeFile(`./photos/teacher/${req.body.image}`, req.file.buffer, (err) => {
-    if (err) return next(err);
-    teacherSchema
-      .updateOne(
-        {
-          _id: req.body._id,
-        },
-        {
-          $set: req.body,
-        }
-      )
-      .then((data) => {
-        if (data.matchedCount === 0) next(new Error("Teacher Not Found"));
-        else res.status(200).json("The Teacher's Data was updated successfully!");
-      })
-      .catch((error) => next(error));
-  });
+  if (req.file) {
+    req.body.image = `${Date.now()}-${Math.random()}-${req.file.originalname}`;
+    fs.writeFile(`./photos/teacher/${req.body.image}`, req.file.buffer, (err) => {
+      if (err) return next(err);
+      updateTeacherData(req, res, next);
+    });
+  } else {
+    updateTeacherData(req, res, next);
+  }
 };
+
+function updateTeacherData(req, res, next) {
+  teacherSchema
+    .updateOne(
+      {
+        _id: req.body._id,
+      },
+      {
+        $set: req.body,
+      }
+    )
+    .then((data) => {
+      if (data.matchedCount === 0) next(new Error("Teacher Not Found"));
+      else res.status(200).json("The Teacher's Data was updated successfully!");
+    })
+    .catch((error) => next(error));
+}
 
 exports.deleteTeacher = (req, res, next) => {
   teacherSchema
@@ -80,13 +97,17 @@ exports.deleteTeacher = (req, res, next) => {
       if (!data) {
         return res.status(404).json({ message: "Teacher not found" });
       }
-      const imagePath = `./photos/teacher/${data.image}`;
-      fs.unlink(imagePath, (err) => {
-        if (err) {
-          return next(err);
-        }
+      if (data.image) {
+        const imagePath = `./photos/teacher/${data.image}`;
+        fs.unlink(imagePath, (err) => {
+          if (err) {
+            return next(err);
+          }
+          res.status(200).json({ message: "Teacher deleted successfully" });
+        });
+      } else {
         res.status(200).json({ message: "Teacher deleted successfully" });
-      });
+      }
     })
     .catch((error) => next(error));
 };
